+++ examples/applySuggestion/index.html
@@ -2,6 +2,11 @@
 <html>
   <head>
     <meta charset="utf-8" />
+
+    <!-- Importmap here simplifies need for bundler in this repo.  if you have a bundler you don't need this. -->
+    <script src="../imports.js"></script>
+    <link rel="stylesheet" href="/style/suggestion-mode.css">
+
     <title>ProseMirror Suggestion Mode Apply Suggestion Demo</title>
     <style>
       body {
