+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "prosemirror-suggestion-mode",
-  "version": "0.1.0",
+  "version": "1.0.68",
   "description": "Track changes and suggestions for ProseMirror",
   "type": "module",
   "main": "dist/index.js",
@@ -30,18 +30,21 @@
     "prosemirror-model": "^1.19.3",
     "@types/jest": "^29.5.0",
     "jest": "^29.5.0",
+    "prosemirror-example-setup": "^1.0.0",
     "jest-environment-jsdom": "^29.5.0",
     "prosemirror-test-builder": "^1.0.0",
     "ts-jest": "^29.1.0",
     "@babel/core": "^7.21.4",
     "@babel/preset-env": "^7.21.4",
-    "@babel/preset-typescript": "^7.21.4"
-  },
+    "@babel/preset-typescript": "^7.21.4",
+    "http-server": "^14.1.1"
+    },
   "scripts": {
     "build": "pm-buildhelper src/index.ts",
-    "watch": "pm-buildhelper src/index.ts --watch",
     "prepare": "npm run build",
     "test": "jest",
-    "test:watch": "jest --watch"
+    "prepublishOnly": "npm test",
+    "start": "http-server . -c-1",
+    "dev": "npm run build & npm run start"
   }
 }
