+++ tsconfig.examples.json
@@ -1,0 +1,22 @@
+{
+  "compilerOptions": {
+    "target": "es2015",
+    "module": "es2015",
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "types": [],
+    "outDir": ".",
+    "rootDir": ".",
+    "baseUrl": ".",
+    "paths": {
+      "prosemirror-suggestion-mode": ["./dist/index.js"],
+      "prosemirror-suggestion-mode/*": ["./dist/*"]
+    }
+  },
+  "include": [
+    "./examples/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
+}
